Move test scaffold to new libvast_test library
