+++ libvast_test/artefacts/logs/bro/dns.log
