+++ libvast_test/artefacts/logs/bro/out.txt
