+++ libvast_test/artefacts/logs/bro/small_conn.log
