+++ libvast_test/artefacts/logs/mrt/bview.20161024.0800
