+++ libvast_test/src/actor_system.cpp
