+++ libvast_test/src/node.cpp
