+++ libvast_test/vast/test/fixtures/filesystem.hpp
